Added outer Makefile

+++ firmware/src/font.c
@@ -152,5 +152,5 @@ uint8_t font_data[FONT_CHAR_COUNT][FONT_CHAR_SIZE] __attribute__((section(".eepr
 };
 
 void font_load(uint8_t* buf, char chr) {
-	eeprom_read_block(buf, font_data[chr], FONT_CHAR_SIZE);
+	eeprom_read_block(buf, font_data[(uint8_t)chr], FONT_CHAR_SIZE);
 }
